build traits
